move exx_helper to module_pwdft
