+++ source/source_pw/module_pwdft/exx_helper.h
@@ -1,6 +1,3 @@
-//
-// For EXX in PW.
-//
 #include "source_psi/psi.h"
 #include "source_base/matrix.h"
 #include "source_pw/module_pwdft/operator_pw/op_exx_pw.h"
