+++ source/source_pw/module_pwdft/hamilt_pw.h
@@ -8,7 +8,7 @@
 #include "source_hamilt/hamilt.h"
 #include "source_pw/module_pwdft/VNL_in_pw.h"
 #include "source_base/kernels/math_kernel_op.h"
-#include "source_pw/module_pwdft/module_exx_helper/exx_helper.h"
+#include "source_pw/module_pwdft/exx_helper.h"
 #include "source_lcao/module_dftu/dftu.h" // mohan add 2025-11-06
 
 namespace hamilt
